Remove unused files, rename files appropriately, move files around for clarity
