+++ client/src/config.js
@@ -1,0 +1,7 @@
+/*
+ * Config
+ */
+
+export default {
+  API_BASE: 'http://localhost:8000'
+}
